+++ crates/da-clients/ethereum/src/lib.rs
@@ -1,8 +1,10 @@
 #![allow(missing_docs)]
 #![allow(clippy::missing_docs_in_private_items)]
 
+use crate::config::EthereumDaConfig;
 use alloy::network::Ethereum;
-use alloy::providers::RootProvider;
+use alloy::providers::{ProviderBuilder, RootProvider};
+use alloy::rpc::client::RpcClient;
 use alloy::transports::http::Http;
 use async_trait::async_trait;
 use color_eyre::Result;
@@ -10,6 +12,12 @@ use da_client_interface::{DaClient, DaVerificationStatus};
 use mockall::automock;
 use mockall::predicate::*;
 use reqwest::Client;
+use std::str::FromStr;
+use url::Url;
+use utils::settings::Settings;
+
+pub const DA_SETTINGS_NAME: &str = "ethereum";
+
 pub mod config;
 pub struct EthereumDaClient {
     #[allow(dead_code)]
@@ -35,5 +43,16 @@ impl DaClient for EthereumDaClient {
 
     async fn max_bytes_per_blob(&self) -> u64 {
         131072
+    }
+}
+
+impl EthereumDaClient {
+    pub fn new_with_settings(settings: &impl Settings) -> Self {
+        let config = EthereumDaConfig::new_with_settings(settings)
+            .expect("Not able to create EthereumDaClient from given settings.");
+        let client =
+            RpcClient::new_http(Url::from_str(config.rpc_url.as_str()).expect("Failed to parse SETTLEMENT_RPC_URL"));
+        let provider = ProviderBuilder::<_, Ethereum>::new().on_client(client);
+        EthereumDaClient { provider }
     }
 }