+++ crates/orchestrator/src/data_storage/aws_s3/config.rs
@@ -1,9 +1,10 @@
-use utils::env_utils::get_env_var_or_panic;
+use serde::{Deserialize, Serialize};
+use utils::settings::Settings;
 
 use crate::data_storage::DataStorageConfig;
 
 /// Represents AWS S3 config struct with all the necessary variables.
-#[derive(Clone)]
+#[derive(Clone, Serialize, Deserialize)]
 pub struct AWSS3Config {
     /// S3 Bucket Name
     pub bucket_name: String,
@@ -12,7 +13,11 @@ pub struct AWSS3Config {
 /// Implementation of `DataStorageConfig` for `AWSS3Config`
 impl DataStorageConfig for AWSS3Config {
     /// To return the config struct by creating it from the environment variables.
-    fn new_from_env() -> Self {
-        Self { bucket_name: get_env_var_or_panic("AWS_S3_BUCKET_NAME") }
+    fn new_with_settings(settings: &impl Settings) -> Self {
+        Self {
+            bucket_name: settings
+                .get_settings("AWS_S3_BUCKET_NAME")
+                .expect("Not able to get AWS_S3_BUCKET_NAME from settings provided."),
+        }
     }
 }