+++ crates/orchestrator/src/data_storage/aws_s3/mod.rs
@@ -1,11 +1,14 @@
+use crate::config::ProviderConfig;
 use crate::data_storage::aws_s3::config::AWSS3Config;
-use crate::data_storage::DataStorage;
+use crate::data_storage::{DataStorage, DataStorageConfig};
 use async_trait::async_trait;
-use aws_config::SdkConfig;
 use aws_sdk_s3::primitives::ByteStream;
 use aws_sdk_s3::Client;
 use bytes::Bytes;
 use color_eyre::Result;
+use utils::settings::Settings;
+
+pub const S3_SETTINGS_NAME: &str = "s3";
 
 /// Module for AWS S3 config structs and implementations
 pub mod config;
@@ -20,16 +23,19 @@ pub struct AWSS3 {
 ///
 /// - initializing a new AWS S3 client
 impl AWSS3 {
-    /// Initializes a new AWS S3 client by passing the config
-    /// and returning it.
-    pub fn new(s3_config: AWSS3Config, aws_config: &SdkConfig) -> Self {
-        // Building AWS S3 config
-        let mut s3_config_builder = aws_sdk_s3::config::Builder::from(aws_config);
-
-        // this is necessary for it to work with localstack in test cases
-        s3_config_builder.set_force_path_style(Some(true));
-        let client = Client::from_conf(s3_config_builder.build());
-        Self { client, bucket: s3_config.bucket_name }
+    /// To init the struct with main settings
+    pub async fn new_with_settings(settings: &impl Settings, provider_config: ProviderConfig) -> Self {
+        match provider_config {
+            ProviderConfig::AWS(aws_config) => {
+                let s3_config = AWSS3Config::new_with_settings(settings);
+                // Building AWS S3 config
+                let mut s3_config_builder = aws_sdk_s3::config::Builder::from(aws_config.as_ref());
+                // this is necessary for it to work with localstack in test cases
+                s3_config_builder.set_force_path_style(Some(true));
+                let client = Client::from_conf(s3_config_builder.build());
+                Self { client, bucket: s3_config.bucket_name }
+            }
+        }
     }
 }