+++ crates/orchestrator/src/database/mongodb/mod.rs
@@ -14,21 +14,26 @@ use mongodb::{
     options::{ClientOptions, ServerApi, ServerApiVersion},
     Client, Collection,
 };
+use utils::settings::Settings;
 use uuid::Uuid;
 
 use crate::database::mongodb::config::MongoDbConfig;
-use crate::database::Database;
+use crate::database::{Database, DatabaseConfig};
 use crate::jobs::types::{JobItem, JobStatus, JobType};
 
 pub mod config;
+
+pub const MONGO_DB_SETTINGS: &str = "mongodb";
 
 pub struct MongoDb {
     client: Client,
 }
 
 impl MongoDb {
-    pub async fn new(config: MongoDbConfig) -> Self {
-        let mut client_options = ClientOptions::parse(config.url).await.expect("Failed to parse MongoDB Url");
+    pub async fn new_with_settings(settings: &impl Settings) -> Self {
+        let mongo_db_settings = MongoDbConfig::new_with_settings(settings);
+        let mut client_options =
+            ClientOptions::parse(mongo_db_settings.url).await.expect("Failed to parse MongoDB Url");
         // Set the server_api field of the client_options object to set the version of the Stable API on the
         // client
         let server_api = ServerApi::builder().version(ServerApiVersion::V1).build();
@@ -39,7 +44,7 @@ impl MongoDb {
         client.database("admin").run_command(doc! {"ping": 1}, None).await.expect("Failed to ping MongoDB deployment");
         println!("Pinged your deployment. You successfully connected to MongoDB!");
 
-        MongoDb { client }
+        Self { client }
     }
 
     /// Mongodb client uses Arc internally, reducing the cost of clone.