+++ crates/orchestrator/src/tests/data_storage/mod.rs
@@ -1,10 +1,12 @@
-use crate::data_storage::aws_s3::config::AWSS3Config;
+use crate::config::{get_aws_config, ProviderConfig};
 use crate::data_storage::aws_s3::AWSS3;
-use crate::data_storage::{DataStorage, DataStorageConfig};
+use crate::data_storage::DataStorage;
 use bytes::Bytes;
 use rstest::rstest;
 use serde_json::json;
+use std::sync::Arc;
 use utils::env_utils::get_env_var_or_panic;
+use utils::settings::env::EnvSettingsProvider;
 
 /// This test checks the ability to put and get data from AWS S3 using `AWSS3`.
 /// It puts JSON data into a test bucket and retrieves it, verifying the data
@@ -14,11 +16,11 @@ use utils::env_utils::get_env_var_or_panic;
 #[tokio::test]
 async fn test_put_and_get_data_s3() -> color_eyre::Result<()> {
     dotenvy::from_filename("../.env.test")?;
-
-    let config = AWSS3Config::new_from_env();
-    let aws_config =
-        aws_config::load_from_env().await.into_builder().endpoint_url(get_env_var_or_panic("AWS_ENDPOINT_URL")).build();
-    let s3_client = AWSS3::new(config, &aws_config);
+    let s3_client = AWSS3::new_with_settings(
+        &EnvSettingsProvider {},
+        ProviderConfig::AWS(Arc::new(get_aws_config(&EnvSettingsProvider {}).await)),
+    )
+    .await;
     s3_client.build_test_bucket(&get_env_var_or_panic("AWS_S3_BUCKET_NAME")).await.unwrap();
 
     let mock_data = json!(