+++ crates/settlement-clients/ethereum/src/lib.rs
@@ -27,10 +27,10 @@ use mockall::{automock, predicate::*};
 
 use alloy::providers::ProviderBuilder;
 use conversion::{get_input_data_for_eip_4844, prepare_sidecar};
-use settlement_client_interface::{SettlementClient, SettlementVerificationStatus, SETTLEMENT_SETTINGS_NAME};
+use settlement_client_interface::{SettlementClient, SettlementConfig, SettlementVerificationStatus};
 #[cfg(test)]
 use url::Url;
-use utils::{env_utils::get_env_var_or_panic, settings::SettingsProvider};
+use utils::env_utils::get_env_var_or_panic;
 
 use crate::clients::interfaces::validity_interface::StarknetValidityContractTrait;
 use crate::clients::StarknetValidityContractClient;
@@ -44,6 +44,7 @@ pub mod conversion;
 mod tests;
 pub mod types;
 
+use utils::settings::Settings;
 use {alloy::providers::RootProvider, alloy::transports::http::Http, reqwest::Client};
 
 pub const ENV_PRIVATE_KEY: &str = "ETHEREUM_PRIVATE_KEY";
@@ -66,9 +67,8 @@ pub struct EthereumSettlementClient {
 }
 
 impl EthereumSettlementClient {
-    pub fn with_settings(settings: &impl SettingsProvider) -> Self {
-        let settlement_cfg: EthereumSettlementConfig = settings.get_settings(SETTLEMENT_SETTINGS_NAME).unwrap();
-
+    pub fn new_with_settings(settings: &impl Settings) -> Self {
+        let settlement_cfg = EthereumSettlementConfig::new_with_settings(settings);
         let private_key = get_env_var_or_panic(ENV_PRIVATE_KEY);
         let signer: PrivateKeySigner = private_key.parse().expect("Failed to parse private key");
         let wallet_address = signer.address();