+++ crates/settlement-clients/settlement-client-interface/src/lib.rs
@@ -2,6 +2,7 @@ use async_trait::async_trait;
 use color_eyre::eyre::Result;
 use mockall::automock;
 use mockall::predicate::*;
+use utils::settings::Settings;
 
 pub const SETTLEMENT_SETTINGS_NAME: &str = "settlement_settings";
 
@@ -52,5 +53,5 @@ pub trait SettlementClient: Send + Sync {
 /// Trait for every new SettlementConfig to implement
 pub trait SettlementConfig {
     /// Should create a new instance of the SettlementConfig from the environment variables
-    fn new_from_env() -> Self;
+    fn new_with_settings(settings: &impl Settings) -> Self;
 }